rename actionCreators to p2pActionCreators and move to p2p directory
